requirements.txt
shotstack-sdk
python-dotenv
requests
